+++ App/Client/Templates/landingPageTemplate.vue
